requirements.txt
ImageHash==4.3.2
pillow==10.4.0